pyproject.toml
[project]
name = "lol-dashboard-v3"
version = "1.0.0"
description = "League of Legends Performance Dashboard"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0",
    "riotwatcher",
    "python-dotenv",
    "pandas",
    "requests",
    "scipy",
    "matplotlib>=3.10.8",
    "jupyter>=1.1.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = []
